updated version code for the ImageGalleryItem component
Changes made:

Converted the ImageGalleryItem component to a functional component by passing the props as an argument.
Removed the incorrect usage of this since it is not required in functional components.
Defined handleClick as a separate function within the component.
Updated the prop types to include the required tags property of the image object.
With these updates, the ImageGalleryItem component should work correctly when used in the ImageGallery component.